requirements.txt
Flask==2.2.3
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.2
Pillow==9.5.0
scikit-learn==1.3.0
